backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "course-creator-backend"
version = "0.1.0"
description = "Backend services for AI-powered course generator"
authors = [{name = "Course Creator Team"}]
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    
    # Task queue
    "celery>=5.3.0",
    "redis>=5.0.0",
    
    # Database & Storage
    "sqlalchemy>=2.0.0",
    "redis>=5.0.0",
    
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    
    # AI & ML
    "dspy-ai>=2.4.0",
    "openai>=1.0.0",
    
    # Data processing
    "pydantic>=2.5.0",
    "python-frontmatter>=1.0.0",
    "gitpython>=3.1.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pathlib>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100 